+++ android/src/main/java/com/transistorsoft/rnbackgroundfetch/RNBackgroundFetchPackage.java
@@ -1,29 +1,44 @@
 package com.transistorsoft.rnbackgroundfetch;
 
-import com.facebook.react.ReactPackage;
-import com.facebook.react.bridge.JavaScriptModule;
+import androidx.annotation.Nullable;
+
+import com.facebook.react.TurboReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
+import com.facebook.react.module.model.ReactModuleInfo;
+import com.facebook.react.module.model.ReactModuleInfoProvider;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
-public class RNBackgroundFetchPackage implements ReactPackage {
-	@Override 
-	public List<NativeModule> createNativeModules (ReactApplicationContext reactContext) {
-        List<NativeModule> modules = new ArrayList<>();
-        modules.add(new RNBackgroundFetchModule(reactContext));
-        return modules;
-    }
+public class RNBackgroundFetchPackage extends TurboReactPackage {
 
-	public List<Class<? extends JavaScriptModule>> createJSModules() {
-        return Collections.emptyList();
-    }
-    
+    @Nullable
     @Override
-    public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-        return Collections.emptyList();
+    public NativeModule getModule(String name, ReactApplicationContext reactContext) {
+        if (RNBackgroundFetchModule.NAME.equals(name)) {
+            return new RNBackgroundFetchModule(reactContext);
+        }
+        return null;
+    }
+
+    @Override
+    public ReactModuleInfoProvider getReactModuleInfoProvider() {
+        return () -> {
+            final Map<String, ReactModuleInfo> map = new HashMap<>();
+            map.put(
+                RNBackgroundFetchModule.NAME,
+                new ReactModuleInfo(
+                    RNBackgroundFetchModule.NAME,
+                    RNBackgroundFetchModule.NAME,
+                    false,  // canOverrideExistingModule
+                    false,  // needsEagerInit
+                    true,   // hasConstants
+                    false,  // isCxxModule
+                    BuildConfig.IS_NEW_ARCHITECTURE_ENABLED
+                )
+            );
+            return map;
+        };
     }
 }